requirements.txt

# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
six

# img
pillow

# web
streamlit
requests
python-dotenv